# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chess Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players/guest:
    post:
      operationId: createGuest
      tags: [players]
      summary: Create a guest player session
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestBody"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSession"
  /players/register:
    post:
      operationId: registerPlayer
      tags: [players]
      summary: Register a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPlayerBody"
      responses:
        "201":
          description: Player registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSession"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/login:
    post:
      operationId: loginPlayer
      tags: [players]
      summary: Login a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPlayerBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/me:
    get:
      operationId: getMe
      tags: [players]
      summary: Get current player profile
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/{playerId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/{playerId}/games:
    get:
      operationId: getPlayerGames
      tags: [players]
      summary: Get player game history
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameListResponse"
  /players/{playerId}/stats:
    get:
      operationId: getPlayerStats
      tags: [players]
      summary: Get player statistics
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
  /players/{playerId}/rating-history:
    get:
      operationId: getPlayerRatingHistory
      tags: [players]
      summary: Get player rating history
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rating history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingPoint"
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [players]
      summary: Get top rated players
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreateGuestBody:
      type: object
      properties:
        username:
          type: string
    RegisterPlayerBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginPlayerBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    PlayerSession:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/Player"
        token:
          type: string
      required:
        - player
        - token
    Player:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        rating:
          type: integer
        isGuest:
          type: boolean
        gamesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - rating
        - isGuest
        - gamesPlayed
        - wins
        - losses
        - draws
        - createdAt
    PlayerStats:
      type: object
      properties:
        playerId:
          type: string
        totalGames:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        winRate:
          type: number
        currentRating:
          type: integer
        peakRating:
          type: integer
        currentStreak:
          type: integer
        favoriteTimeControl:
          type: ["string", "null"]
        avgGameDuration:
          type: ["number", "null"]
      required:
        - playerId
        - totalGames
        - wins
        - losses
        - draws
        - winRate
        - currentRating
        - peakRating
        - currentStreak
        - favoriteTimeControl
        - avgGameDuration
    RatingPoint:
      type: object
      properties:
        rating:
          type: integer
        gameId:
          type: string
        timestamp:
          type: string
          format: date-time
        ratingChange:
          type: integer
      required:
        - rating
        - gameId
        - timestamp
        - ratingChange
    TimeControl:
      type: object
      properties:
        initialSeconds:
          type: integer
        incrementSeconds:
          type: integer
        label:
          type: string
      required:
        - initialSeconds
        - incrementSeconds
        - label
    Game:
      type: object
      properties:
        id:
          type: string
        whitePlayerId:
          type: string
        blackPlayerId:
          type: ["string", "null"]
        whiteUsername:
          type: string
        blackUsername:
          type: ["string", "null"]
        whiteRating:
          type: integer
        blackRating:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, active, finished, aborted]
        result:
          type: ["string", "null"]
          enum: [white, black, draw, null]
        resultReason:
          type: ["string", "null"]
        fen:
          type: string
        pgn:
          type: string
        timeControl:
          $ref: "#/components/schemas/TimeControl"
        whiteTimeMs:
          type: integer
        blackTimeMs:
          type: integer
        isAiGame:
          type: boolean
        aiDepth:
          type: ["integer", "null"]
        roomCode:
          type: ["string", "null"]
        drawOfferedBy:
          type: ["string", "null"]
        moveCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - whitePlayerId
        - blackPlayerId
        - whiteUsername
        - blackUsername
        - whiteRating
        - blackRating
        - status
        - result
        - resultReason
        - fen
        - pgn
        - timeControl
        - whiteTimeMs
        - blackTimeMs
        - isAiGame
        - aiDepth
        - roomCode
        - drawOfferedBy
        - moveCount
        - createdAt
        - updatedAt
...
    GameListResponse:
      type: object
      properties:
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total:
          type: integer
      required:
        - games
        - total
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        playerId:
          type: string
        username:
          type: string
        rating:
          type: integer
        gamesPlayed:
          type: integer
        winRate:
          type: number
      required:
        - rank
        - playerId
        - username
        - rating
        - gamesPlayed
        - winRate
...